Added Classification metrics, operators and functions
Classification metrics added for fitness funtion:
Weighted Accuracy
Recall
Precision
Calculate harmonic mean of the classes, and for 2 clases 0 is the goal.

Added "<", "<=", ">", ">=", "&&", "||" as operators for the grammars.

Added function ifElse( condition, true, false) where the first argument is the condition, the second is what returns if true and the third what returns if false.